Set up potions for easy data retrieval

Start-up function to place a list into options with a specified slot name.

brew(..., file, .slot, .pkg)

brew_package(..., file, .pkg)

brew_interactive(..., file, .slot)

Arguments

...

One or more == statements giving attributes to be stored; or alternatively a list containing the same.

file

string: optional file containing data to be stored via options(). Valid formats are .yml or .json.

.slot

string: optional name to mandate where data is stored. Defaults to a random string generated by stringi::stri_rand_strings.

.pkg

string: package name that {potions} is being used within. If using potions within a package development process, set this argument instead of slot.

Details

The default method is to use brew and set either .pkg or .slot, but not both. Alternatively you can use brew_package() or brew_interactive(). Note that if neither .slot or .pkg are set, potions defaults to .slot , unless .pkg information has previously been supplied (and .slot information has not). This might be undesirable in a package development situation.

If both ... and file arguments are empty, this function sets up an empty potions object in options("potions-pkg"); See potions-class for more information on this data type. If ... and file arguments are provided, they will be amalgamated. If there are identical names in both lists, those in ... are chosen.

If the user repeatedly calls brew(), later list entries overwrite early entries. Whole lists are not overwritten unless all top-level entry names match.

Examples

# basic usage is to pass expressions
brew(x == 1)

# lists are also permitted
brew(list(x = 2))

# as are passing lists as objects
my_list <- list(x = 3)
brew(my_list)

# optional clean-up
drain()